// apps/organization/src/app/projects/enums/ProjectEnum.ts
export enum ProjectEnum {
  PROJECTS = 'Projects',
  NEW_PROJECT = 'New Project',
  DEPARTMENT = 'Department',
  CLIENT = 'Client',
  PROJECT = 'Project',
  MANAGER = 'Manager',
  SHOWING = 'Showing',
  ENTRIES = 'entries',
  PREVIOUS = 'Previous',
  NEXT = 'Next',
  NAME = 'name',
  CODE = 'code',
  ASC = 'ASC',
  DESC = 'DESC',
  PROJECT_NAME = 'Name',
  CREATED_AT = 'createdAt',
  ACTION = 'Action',
  NO_RECORDS = 'No records...',
}

export enum AddProjectEnum {
  ADD_NEW_PROJECT = 'Add New Projects',
  BACK = 'Back',
  PROJECT_NAME = 'Project Name',
  CODE = 'code',
  PROJECT_Key_CODE = 'Project Key/Code',
  PROJECT_CODE = ' Key/Code',
  DEPARTMENT_ID = 'departmentIds',
  DEPARTMENT_LABEL = 'department-label',
  CLIENT_ID = 'clientId',
  START_DATE = 'startDate',
  START_DATE_LABEL = 'Start Date',
  END_DATE = 'endDate',
  END_DATE_LABEL = 'End Date',
  DESCRIPTION = 'description',
  DESCRIPTION_LABEL = 'Description',
  BUDGETED_HOURS = 'Budgeted Hours',
  BUDGETED_HOURS_NAME = 'budgetedHours',
  CANCEL = 'Cancel',
  UPDATE = 'Update',
  SAVE = 'Save',
  EDIT_PROJECT = 'Edit Project',
}